+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "GearScout",
-  "name": "GearScout",
+  "short_name": "QualScout",
+  "name": "QualScout",
   "icons": [
     {
       "src": "logos/32-favicon.png",
