+++ src/components/data-collection-page/match-information/MatchInformation.scss
