+++ src/components/data-collection-page/qualitative-section/QualitativePage.scss
