--- src/index.css
@@ -1,21 +1,0 @@
-body {
-  margin: 0;
-  font-family: Roboto, 'Segoe UI', -apple-system, BlinkMacSystemFont, 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-/*
-@font-face {
-  font-family: 'Segoe UI';
-  src: url('src\fonts\Segoe UI.ttf') format('woff'),
-       url('src\fonts\Segoe UI.woff') format('truetype');
-}
-*/
