+++ src/index.scss
@@ -1,0 +1,13 @@
+body {
+	margin: 0;
+	font-family: Roboto, 'Segoe UI', -apple-system, BlinkMacSystemFont, 'Oxygen',
+		'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+		sans-serif;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+}
+
+code {
+	font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+		monospace;
+}
