+++ src/index.tsx
@@ -2,9 +2,9 @@ import { createTheme, ThemeProvider } from '@mui/material/styles';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import App from './App';
-import store from './app/Store';
-import './index.css';
+import App from './components/App';
+import store from './state/Store';
+import './index.scss';
 
 export const themeOptions = {
 	palette: {
